requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
httpx==0.24.1
streamlit==1.25.0
pydantic==1.10.2
requests==2.28.1
python-dotenv==1.0.0
pydantic-settings==2.1.0
aiohttp==3.9.1
tenacity==8.2.3
python-logging-loki==0.3.1
google-adk==0.5.0
python-a2a==0.3.0
langchain==0.1.0
sentence-transformers==2.2.2